+++ src/App.js
@@ -1,10 +1,12 @@
 import "./App.css";
 import Header from "./components/Layout/Header";
+import Meals from "./components/Meal/Meals";
 
 function App() {
 	return (
 		<>
 			<Header />
+			<Meals />
 		</>
 	);
 }