+++ src/components/Meal/Meals.js
@@ -1,0 +1,7 @@
+import MealsTitle from "./MealsTitle";
+
+const Meals = () => {
+	return <MealsTitle />;
+};
+
+export default Meals;